execução de tela de login

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\LoginController;
 use App\Http\Controllers\ScoreController;
 use Illuminate\Support\Facades\Route;
 
@@ -17,6 +18,10 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [ScoreController::class, 'index'])->name('index-score');
 Route::get('/add', [ScoreController::class, 'create'])->name('add-score');
 Route::post('/add', [ScoreController::class, 'store']);
-Route::get('/admin', [ScoreController::class, 'admin'])->name('admin');
-Route::get('/destroy/{id}', [ScoreController::class, 'destroyShow'])->name('delete-confirmation');
-Route::delete('/destroy{id}', [ScoreController::class, 'destroy'])->name('delete');
+Route::get('/admin', [ScoreController::class, 'admin'])->name('admin')->middleware('auth');
+Route::get('/destroy/{id}', [ScoreController::class, 'destroyShow'])->name('delete-confirmation')->middleware('auth');
+Route::delete('/destroy{id}', [ScoreController::class, 'destroy'])->name('delete')->middleware('auth');
+
+Route::get('/login', [LoginController::class, 'login'])->name('login');
+Route::post('/login', [LoginController::class, 'authenticate']);
+Route::post('/logout', [LoginController::class, 'logout'])->name('logout');
